"Actions" and "Filters" classes renamed and refactored.
